+++ core/forms.py
@@ -1,5 +1,21 @@
-from django import forms
+from django import forms as django_forms
+from directory_components import forms
 
 
-class CompaniesHouseSearchForm(forms.Form):
-    term = forms.CharField()
+class CompaniesHouseSearchForm(django_forms.Form):
+    term = django_forms.CharField()
+
+
+class SectorPotentialForm(forms.Form):
+
+    SECTOR_CHOICES_BASE = [('', 'To see possible markets - select your sector')]
+
+    sector = forms.ChoiceField(
+        label='Sector',
+        choices=SECTOR_CHOICES_BASE,
+    )
+
+    def __init__(self, sector_list, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        sorted_sectors = sorted(sector_list, key=lambda x: x['name'])
+        self.fields['sector'].choices = self.SECTOR_CHOICES_BASE + [(tag['id'], tag['name']) for tag in sorted_sectors]