+++ core/views.py
@@ -10,7 +10,7 @@ from directory_forms_api_client.helpers import FormSessionMixin, Sender
 from django.conf import settings
 from django.contrib import sitemaps
 from django.http import JsonResponse
-from django.urls import reverse, RegexURLResolver
+from django.urls import reverse
 from django.utils.cache import set_response_etag
 from django.views.generic import FormView, TemplateView
 from django.views.generic.base import RedirectView, View
@@ -37,13 +37,17 @@ class LandingPageView(mixins.SetGA360ValuesForCMSPageMixin, TemplateView):
     @property
     def template_name(self):
         alt_landing_page_requested = self.request.GET.get('nh')
-        required_cms_fields = ['how_dit_helps_title', 'hero_text', 'questions_section_title', 'what_is_new_title']
+        required_cms_fields = ['how_dit_helps_title', 'hero_text', 'what_is_new_title']
         cms_page_has_new_fields = all([key in self.page for key in required_cms_fields])
 
         if cms_page_has_new_fields and alt_landing_page_requested:
             return 'core/landing_page_alternate.html'
 
         return 'core/landing_page_domestic.html'
+
+    @cached_property
+    def sector_list(self):
+        return helpers.handle_cms_response(cms_api_client.list_industry_tags())
 
     @cached_property
     def page(self):
@@ -60,8 +64,12 @@ class LandingPageView(mixins.SetGA360ValuesForCMSPageMixin, TemplateView):
         return super().get(request, *args, **kwargs)
 
     def get_context_data(self, *args, **kwargs):
+        sorted_sectors = sorted(self.sector_list, key=lambda x: x['pages_count'], reverse=True)
+        top_sectors = sorted_sectors[:6]
         return super().get_context_data(
             page=self.page,
+            sector_list=top_sectors,
+            sector_form=forms.SectorPotentialForm(sector_list=self.sector_list),
             casestudies=[
                 casestudies.HELLO_BABY,
                 casestudies.YORK,
@@ -192,8 +200,7 @@ class StaticViewSitemap(sitemaps.Sitemap):
 
         return [
             item.name for item in urls.urlpatterns
-            if not isinstance(item, RegexURLResolver) and
-            item not in redirects and
+            if item not in redirects and
             item.name not in excluded_pages
         ]